open ai chat box/backend_python/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
openai>=1.0.0
openai
python-dotenv
requests==2.31.0